+++ include/jsoncons_ext/bson/bson_reader.hpp
@@ -275,8 +275,11 @@ private:
 };
 
 typedef basic_bson_reader<jsoncons::binary_stream_source> bson_reader;
+typedef basic_bson_reader<jsoncons::bytes_source> bson_bytes_reader;
 
-typedef basic_bson_reader<jsoncons::byte_array_source> bson_buffer_reader;
+#if !defined(JSONCONS_NO_DEPRECATED)
+typedef bson_bytes_reader bson_buffer_reader;
+#endif
 
 }}