+++ include/jsoncons_ext/cbor/cbor_reader.hpp
@@ -1347,7 +1347,11 @@ private:
 
 typedef basic_cbor_reader<jsoncons::binary_stream_source> cbor_reader;
 
-typedef basic_cbor_reader<jsoncons::byte_array_source> cbor_buffer_reader;
+typedef basic_cbor_reader<jsoncons::bytes_source> cbor_bytes_reader;
+
+#if !defined(JSONCONS_NO_DEPRECATED)
+typedef cbor_bytes_reader cbor_buffer_reader;
+#endif
 
 }}