+++ tests/src/ubjson/encode_ubjson_tests.cpp
@@ -137,7 +137,7 @@ TEST_CASE("encode_ubjson_arrays_and_maps")
 TEST_CASE("encode indefinite length ubjson arrays and maps")
 {
     std::vector<uint8_t> v;
-    ubjson_buffer_encoder encoder(v);
+    ubjson_bytes_encoder encoder(v);
 
     SECTION("[\"Hello\"]")
     {